Add controller layer
Add create and get products API's
Add virtual attribute for Products
