+++ backend/src/controllers/index.js
@@ -1,0 +1,2 @@
+export * as authController from './authController.js';
+export * as productController from './productController.js';
