+++ backend/src/routes/auth.js
@@ -1,22 +1,9 @@
 import express from 'express';
 
-import { authServices } from '../services/index.js';
+import { authController } from '../controllers/index.js';
 
 const authRouter = express.Router();
 
-authRouter.post('/login', async (req, res) => {
-  try {
-    const { type, email, password, refreshToken } = req.body;
-    const signedUser = await authServices.login({
-      type,
-      email,
-      password,
-      refreshToken,
-    });
-    res.status(200).json(signedUser);
-  } catch (error) {
-    res.status(401).json({ message: error.message });
-  }
-});
+authRouter.post('/login', authController.login);
 
 export { authRouter };
